fix: work around WeChat WebGL VAO state restoration bug
WeChat's WebGL implementation doesn't correctly store/restore
per-attribute VBO references in VAOs. Explicitly rebind VBO,
vertex attribute pointers and IBO before every draw call.

- BatchRenderer2D::flush(): bypass VAO, set attrib pointers directly
- CustomGeometry::bind(): rebind VBO + attribs + IBO after VAO bind
- draw_mesh/draw_meshWithUniforms: call glDrawElements directly
- Index buffers u32 → u16 for WebGL compatibility (batch, spine, quad)
- Batch shader: add GLSL ES 1.0 fallback, switch → if-else chain
- Sampler uniforms moved from per-flush to per-init
- Clean up all frame-level diagnostic logging

+++ src/esengine/renderer/RenderContext.cpp
@@ -66,7 +66,7 @@ void RenderContext::initQuadData() {
         -0.5f,  0.5f,       0.0f, 1.0f
     };
 
-    u32 indices[] = { 0, 1, 2, 2, 3, 0 };
+    u16 indices[] = { 0, 1, 2, 2, 3, 0 };
 
     quadVAO_ = VertexArray::create();
